move test to index_sorter
